Lockpoint interpolation and min. pedal
Implemented interpolation between two lockpoints and added speed and target lock readouts to the App.

Also added a minimum pedal position in order for 50/50 and custom modes to work - settable between 0 and 100%. This is to prevent wheels skipping when parking or performing slow operations - 5 to 10% will do the job just fine I think.

Still untested.....